Fixed bundle addCommand behaviour
* Changed BaseKernel place. Keept old reference to maintain BC
