+++ Tests/BaseKernel.php
@@ -15,179 +15,11 @@ declare(strict_types=1);
 
 namespace Mmoreram\BaseBundle\Tests;
 
-use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
-use Symfony\Component\Config\Loader\LoaderInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\HttpKernel\Bundle\BundleInterface;
-use Symfony\Component\HttpKernel\Kernel;
-use Symfony\Component\Routing\RouteCollectionBuilder;
-use Symfony\Component\Yaml\Yaml;
-
-use Mmoreram\SymfonyBundleDependencies\BundleDependenciesResolver;
+use Mmoreram\BaseBundle\Kernel\BaseKernel as MainBaseKernel;
 
 /**
  * Class BaseKernel.
  */
-final class BaseKernel extends Kernel
+final class BaseKernel extends MainBaseKernel
 {
-    use MicroKernelTrait;
-    use BundleDependenciesResolver;
-
-    /**
-     * @var string[]
-     *
-     * Bundle array
-     */
-    private $bundlesToLoad;
-
-    /**
-     * @var array[]
-     *
-     * Configuration
-     */
-    private $configuration;
-
-    /**
-     * @var array[]
-     *
-     * Routes
-     */
-    private $routes;
-
-    /**
-     * BaseKernel constructor.
-     *
-     * @param string[] $bundlesToLoad
-     * @param array[]  $configuration
-     * @param array[]  $routes
-     */
-    public function __construct(
-        array $bundlesToLoad,
-        array $configuration = [],
-        array $routes = []
-    ) {
-        $this->bundlesToLoad = $bundlesToLoad;
-        $this->routes = $routes;
-        $this->configuration = array_merge(
-            [
-                'parameters' => [
-                    'kernel.secret' => '1234',
-                ],
-            ],
-            $configuration
-        );
-
-        parent::__construct('test', false);
-    }
-
-    /**
-     * Returns an array of bundles to register.
-     *
-     * @return BundleInterface[] An array of bundle instances
-     */
-    public function registerBundles()
-    {
-        return $this->getBundleInstances(
-            $this,
-            $this->bundlesToLoad
-        );
-    }
-
-    /**
-     * Configures the container.
-     *
-     * You can register extensions:
-     *
-     * $c->loadFromExtension('framework', array(
-     *     'secret' => '%secret%'
-     * ));
-     *
-     * Or services:
-     *
-     * $c->register('halloween', 'FooBundle\HalloweenProvider');
-     *
-     * Or parameters:
-     *
-     * $c->setParameter('halloween', 'lot of fun');
-     *
-     * @param ContainerBuilder $c
-     * @param LoaderInterface  $loader
-     */
-    protected function configureContainer(
-        ContainerBuilder $c,
-        LoaderInterface $loader
-    ) {
-        $yamlContent = Yaml::dump($this->configuration);
-        $filePath = sys_get_temp_dir() . '/base-test-' . rand(1, 9999999) . '.yml';
-        file_put_contents($filePath, $yamlContent);
-        $loader->load($filePath);
-        unlink($filePath);
-    }
-
-    /**
-     * Add or import routes into your application.
-     *
-     *     $routes->import('config/routing.yml');
-     *     $routes->add('/admin', 'AppBundle:Admin:dashboard', 'admin_dashboard');
-     *
-     * @param RouteCollectionBuilder $routes
-     */
-    protected function configureRoutes(RouteCollectionBuilder $routes)
-    {
-        foreach ($this->routes as $route) {
-            is_array($route)
-                ? $routes->add(
-                    $route[0],
-                    $route[1],
-                    $route[2]
-                )
-                : $routes->import($route);
-        }
-    }
-
-    /**
-     * Gets the application root dir.
-     *
-     * @return string The application root dir
-     */
-    public function getRootDir()
-    {
-        $bundles = array_map(function ($bundle) {
-            return is_object($bundle)
-                ? get_class($bundle)
-                : $bundle;
-        }, $this->bundlesToLoad);
-        $config = $this->configuration;
-        $routes = $this->routes;
-        sort($bundles);
-        sort($routes);
-        $this->sortArray($config);
-
-        return sys_get_temp_dir() . '/base-kernel/' . 'kernel-' . (
-            hash(
-                'md5',
-                json_encode([
-                    $bundles,
-                    $config,
-                    $routes,
-                ])
-            )
-        );
-    }
-
-    /**
-     * Sort array's first level, taking in account if associative array or
-     * sequential array.
-     *
-     * @param mixed $element
-     */
-    private function sortArray(&$element)
-    {
-        if (is_array($element)) {
-            array_walk($element, [$this, 'sortArray']);
-            array_key_exists(0, $element)
-                ? sort($element)
-                : ksort($element);
-        }
-    }
 }
